feat: add kyc validation
