+++ contracts/scripts/constants.ts
@@ -512,7 +512,7 @@ export const GAS_LIMIT = {
     businessLogicResolver: {
         getStaticResolverKey: 60_000,
         registerBusinessLogics: 7_800_000,
-        createConfiguration: 20_000_000,
+        createConfiguration: 15_000_000,
     },
 }
