+++ sdk/__tests__/utils/verifiableCredentials.ts
@@ -203,44 +203,45 @@
 
 */
 
-import Account from './Account.js';
-import Role from './Role.js';
-import Security from './Security.js';
-import Equity from './Equity.js';
-import Bond from './Bond.js';
-import Event from './Event.js';
-import Network from './Network.js';
-import Factory from './Factory.js';
-import Management from './Management.js';
-import SSIManagement from './SSIManagement.js';
-import Kyc from './Kyc.js';
+import { ethers } from 'ethers';
+import { createEcdsaCredential, EthrDID } from '@terminal3/ecdsa_vc';
+import { DID, type VerificationOptions } from '@terminal3/vc_core';
+import PrivateKey from '../../src/domain/context/account/PrivateKey';
+import { CLIENT_ACCOUNT_ECDSA } from '../config';
 
-export {
-  Security,
-  Equity,
-  Bond,
-  Account,
-  Role,
-  Event,
-  Network,
-  Factory,
-  Management,
-  SSIManagement,
-  Kyc
-};
+async function createVcT3(address: string): Promise<string> {
+  const issuerPrivateKey = CLIENT_ACCOUNT_ECDSA.privateKey as PrivateKey;
+  const issuer = new EthrDID(issuerPrivateKey.key as string, 'polygon');
+  const holderDid = new DID('ethr', address);
 
-export * from './request';
-export * from './response';
+  // Creating a credential with BBS+ signature
+  const claims = { kyc: 'passed' };
+  const revocationRegistryAddress =
+    '0x77Fb69B24e4C659CE03fB129c19Ad591374C349e';
+  const didRegistryAddress = '0x312C15922c22B60f5557bAa1A85F2CdA4891C39a';
+  const provider = new ethers.providers.JsonRpcProvider(
+    'https://testnet.hashio.io/api',
+  );
+  const options = {
+    revocationRegistryAddress,
+    provider,
+    didRegistryAddress,
+  } as unknown as VerificationOptions;
 
-export * from './Security.js';
-export * from './Equity.js';
-export * from './Bond.js';
-export * from './Account.js';
-export * from './Role.js';
-export * from './Event.js';
-export * from './Common.js';
-export * from './Network.js';
-export * from './Factory.js';
-export * from './Management.js';
-export * from './Kyc.js';
-export * from './SSIManagement.js';
+  const vc = await createEcdsaCredential(
+    issuer,
+    holderDid,
+    claims,
+    ['KycCredential'],
+    undefined,
+    undefined,
+    options,
+  );
+
+  const vcString = JSON.stringify(vc);
+  const vcBase64 = Buffer.from(vcString).toString('base64');
+
+  return vcBase64;
+}
+
+export default createVcT3;
