+++ sdk/src/app/usecase/command/security/error/AccountAlreadyIsAnIssuer.ts
